more clean bar code
