+++ RES/src/app/dashboard/components/bar/bar.component.html
@@ -9,27 +9,27 @@
   (expanded)="cms.disPatchSetInView($event)"
   (chartInstance)="handleChartInstance($event)"
 >
-  <ng-container *ngIf="buckets">
+  <ng-container *ngIf="barService.buckets">
     <div class="col-12">
       <ng-select
-        [items]="buckets.type"
+        [items]="barService.buckets.type"
         bindLabel="key"
         [multiple]="true"
         [dropdownPosition]="'top'"
         placeholder="categories"
-        [(ngModel)]="selectedCategories"
+        [(ngModel)]="barService.selectedCategories"
       >
       </ng-select>
     </div>
 
     <div class="col-12">
       <ng-select
-        [items]="buckets['year.keyword']"
+        [items]="barService.buckets['year.keyword']"
         bindLabel="key"
         [multiple]="true"
         [dropdownPosition]="'top'"
         placeholder="years"
-        [(ngModel)]="selectedYears"
+        [(ngModel)]="barService.selectedYears"
       >
       </ng-select>
     </div>
