+++ RES/src/app/dashboard/components/bar/services/bar/bar.service.ts
@@ -1,10 +1,41 @@
-import { Injectable } from '@angular/core';
+import { Injectable, EventEmitter } from '@angular/core';
 import { BodyBuilderService } from 'src/app/filters/services/bodyBuilder/body-builder.service';
 import * as bodybuilder from 'bodybuilder';
-import { ElasticsearchQuery } from 'src/app/filters/services/interfaces';
+import {
+  ElasticsearchQuery,
+  ResetOptions,
+  ElasticsearchResponse,
+  BucketWithInnerBuckts,
+  Bucket,
+} from 'src/app/filters/services/interfaces';
+import { MergedSelect } from 'src/configs/generalConfig.interface';
+import { Subject } from 'rxjs';
+import { ItemsService } from 'src/services/itemsService/items.service';
+import { RangeService } from 'src/app/filters/services/range/range.service';
+import { map, first } from 'rxjs/operators';
+import { ESHttpError } from 'src/store/actions/actions.interfaces';
+
+interface ComposerHelper {
+  subFlag: boolean;
+  forgetYearsFromStore: boolean;
+}
 @Injectable()
 export class BarService {
-  constructor(private readonly bodyBuilderService: BodyBuilderService) {}
+  buckets: MergedSelect;
+  setChartOptinos: EventEmitter<Array<Highcharts.SeriesColumnOptions>>;
+  selectedCategories: Array<Bucket>;
+  selectedYears: Array<Bucket>;
+  private rangeService: RangeService;
+
+  constructor(private readonly itemsService: ItemsService) {
+    this.setChartOptinos = new EventEmitter<
+      Array<Highcharts.SeriesColumnOptions>
+    >();
+  }
+
+  init(rangeService: RangeService) {
+    this.rangeService = rangeService;
+  }
 
   buildQuery(queryToMerge?: ElasticsearchQuery): ElasticsearchQuery {
     const finalQuery: ElasticsearchQuery = {
@@ -28,5 +59,103 @@ export class BarService {
       delete finalQuery.query;
     }
     return finalQuery;
+  }
+
+  yearsComposer(source: string): (b: MergedSelect) => void {
+    const composer: ComposerHelper = {
+      subFlag: true,
+      forgetYearsFromStore: false,
+    };
+    return (b: MergedSelect) => {
+      if (composer.subFlag) {
+        this.getData();
+        this.subToShouldReset(composer, source);
+      }
+      if (!composer.forgetYearsFromStore) {
+        this.buckets = { ...b };
+      }
+      composer.subFlag = false;
+      composer.forgetYearsFromStore = false;
+    };
+  }
+
+  private subToShouldReset(composer: ComposerHelper, source: string) {
+    this.rangeService.shouldReset.subscribe(
+      ({ caller, data }: ResetOptions) => {
+        if (caller === 'range' && data) {
+          const { min, max } = data;
+          const filterdYearsRange = [];
+          for (let i = min; i < max; i++) {
+            filterdYearsRange.push(i);
+          }
+          this.buckets = {
+            ...this.buckets,
+            [source]: filterdYearsRange,
+          };
+          composer.forgetYearsFromStore = true;
+        }
+        this.getData();
+      }
+    );
+  }
+
+  private getData(): void {
+    this.itemsService
+      .getItems(
+        this.buildQuery(this.rangeService
+          .buildquery({ size: 12 })
+          .build() as ElasticsearchQuery)
+      )
+      .pipe(
+        map(this.mapDataToColmns.bind(this)),
+        first()
+      )
+      .subscribe(
+        (series: Array<Highcharts.SeriesColumnOptions>) => {
+          this.setChartOptinos.emit(series);
+        },
+        (error: ESHttpError) => this.setChartOptinos.emit()
+      );
+  }
+  private mapDataToColmns(
+    res: ElasticsearchResponse
+  ): Array<Highcharts.SeriesColumnOptions> {
+    const series: Array<
+      Highcharts.SeriesColumnOptions
+    > = res.aggregations.y.buckets.map(
+      (yBucket: BucketWithInnerBuckts) =>
+        ({
+          type: 'column',
+          name: yBucket.key,
+          value: yBucket.doc_count,
+          data: yBucket.x.buckets.map(xBucket => ({
+            name: xBucket.key,
+            y: xBucket.doc_count,
+          })),
+        } as Highcharts.SeriesColumnOptions)
+    );
+    this.selectDefaultOptions(series);
+    return series;
+  }
+
+  private selectDefaultOptions(
+    series: Array<Highcharts.SeriesColumnOptions>
+  ): void {
+    const [{ data }] = series;
+
+    this.selectedCategories = data.map(
+      ({ y, name }: { y: number; name: string }) =>
+        ({
+          key: name,
+          doc_count: y,
+        } as Bucket)
+    );
+    this.selectedYears = series.map(
+      ({ name, value }: Highcharts.SeriesColumnOptions & { value: any }) =>
+        ({
+          key: name,
+          doc_count: value,
+        } as Bucket)
+    );
   }
 }
